Update Audio Player
Added how to use html5 audio tag (somewhat) in c# on server-side. As in, how to change it's source.
Still have to find out how to automatically play the newly selected song so the user doesn't have to press the button themself each time they click on a song in a playlist.